onefin_project/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.19
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
factory-boy==3.3.3
Faker==36.1.1
idna==3.10
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
